add validations in sigup

+++ maazweb/Client/user/src/component/molecules/login-model/signup/Signup.js
@@ -2,8 +2,7 @@ import React, { useState } from "react";
 import styles from "./signup.module.css";
 import axios from "axios";
 import Swal from "sweetalert2";
-import * as validate from "../../../../utils/validations/Validations";
-import { API_URL } from "../../../constant/WebsiteConstants";
+import { API_URL, WEBSITE_NAME } from "../../../constant/WebsiteConstants";
 import OTP from "../opt-verification/OTP";
 import Button from "../../../atoms/button/Button";
 
@@ -20,80 +19,55 @@ function Signup({ onClose }) {
   const [showOtpForm, setShowOtpForm] = useState(false);
   const [email, setEmail] = useState("");
 
-  const handleChange = (e) => {
-    const { name, value, type, checked } = e.target;
-    if (type === "checkbox") {
-      setIsChecked(checked);
-    } else {
-      switch (name) {
-        case "userName":
-        case "country":
-          if (!validate.name(value)) {
-            setError(`${name} cannot have special characters or digits`);
-          } else {
-            setError("");
-          }
-          break;
-        case "email":
-          if (!validate.email(value)) {
-            setError("Email must fulfill requirements");
-          } else {
-            setError("");
-          }
-          break;
-        case "password":
-        case "confirmPassword":
-          if (!validate.password(value)) {
-            setError(`${name} must fulfill requirements`);
-          } else {
-            setError("");
-          }
-          break;
-        default:
-          break;
-      }
-      setFormData((prevFormData) => ({ ...prevFormData, [name]: value }));
-    }
-  };
-
   const handleSubmit = async (e) => {
     e.preventDefault();
-    if (formData.password !== formData.confirmPassword) {
+
+    // Check if required fields are filled
+    const { userName, country, email, password, confirmPassword } = formData;
+
+    if (!userName || !country || !email || !password || !confirmPassword || !isChecked) {
+      setError(true);
+      return;
+    }
+    setError(false);
+
+    if (password !== confirmPassword) {
       setError("Passwords do not match");
       return;
     }
 
-    if (!isChecked) {
-      setError("You must consent to the processing of your personal information");
-      return;
-    }
-  
     try {
       setError("");
       const response = await axios.post(`${API_URL}/api/user/user_signUp`, formData);
       console.log("API response:", response.data);
-  
+
       // Assuming the token is in response.data.token
       const token = response.data.token;
-  
+
       // Store the token in local storage
       localStorage.setItem("access_token", token);
-  
+
       setEmail(formData.email);
       setShowOtpForm(true);
     } catch (error) {
-      console.error("Error:", error.response ? error.response.data : error.message);
       Swal.fire("Sign Up Failed!", "Please check your information and try again.", "error");
       setError("An error occurred. Please try again.");
     }
   };
-  
 
   const closeOtpForm = () => {
     setShowOtpForm(false);
     onClose();
   };
 
+  // Handle form input changes
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setFormData((prevFormData) => ({
+      ...prevFormData,
+      [name]: value,
+    }));
+  };
 
   return (
     <>
@@ -109,74 +83,111 @@ function Signup({ onClose }) {
 
             <form onSubmit={handleSubmit} className={styles.form}>
               <div className={styles.formRow}>
-                <input
-                  onChange={handleChange}
-                  type="text"
-                  name="userName"
-                  value={formData.userName}
-                  placeholder="UserName"
-                />
-                <input
-                  onChange={handleChange}
-                  type="text"
-                  name="country"
-                  value={formData.country}
-                  placeholder="Country"
-                />
+                <div className={styles.colItem}>
+                  <input
+                    type="text"
+                    name="userName"
+                    value={formData.userName}
+                    onChange={handleChange}
+                    placeholder="UserName"
+                  />
+                  {error && !formData.userName && (
+                    <span className={styles.text_danger}>User Name is required</span>
+                  )}
+                </div>
+
+                <div className={styles.colItem}>
+                  <input
+                    type="text"
+                    name="country"
+                    value={formData.country}
+                    onChange={handleChange}
+                    placeholder="Country"
+                  />
+                  {error && !formData.country && (
+                    <span className={styles.text_danger}>Country is required</span>
+                  )}
+                </div>
               </div>
 
               <div className={styles.formRow}>
-                <input
-                  onChange={handleChange}
-                  type="email"
-                  name="email"
-                  value={formData.email}
-                  placeholder="Email Address"
-                />
+                <div className={styles.colMail}>
+                  <input
+                    type="email"
+                    name="email"
+                    value={formData.email}
+                    onChange={handleChange}
+                    placeholder="Email Address"
+                  />
+                  {error && !formData.email && (
+                    <span className={styles.text_danger}>Email is required</span>
+                  )}
+                </div>
               </div>
 
               <div className={styles.formRow}>
-                <input
-                  onChange={handleChange}
-                  type="password"
-                  name="password"
-                  value={formData.password}
-                  placeholder="Password"
-                />
-                <input
-                  onChange={handleChange}
-                  type="password"
-                  name="confirmPassword"
-                  value={formData.confirmPassword}
-                  placeholder="Confirm Password"
-                />
+                <div className={styles.colItem}>
+                  <input
+                    type="password"
+                    name="password"
+                    value={formData.password}
+                    onChange={handleChange}
+                    placeholder="Password"
+                  />
+                  {error && !formData.password && (
+                    <span className={styles.text_danger}>Password is required</span>
+                  )}
+                </div>
+                <div className={styles.colItem}>
+                  <input
+                    type="password"
+                    name="confirmPassword"
+                    value={formData.confirmPassword}
+                    onChange={handleChange}
+                    placeholder="Confirm Password"
+                  />
+                  {error && !formData.confirmPassword && (
+                    <span className={styles.text_danger}>Confirm Password is required</span>
+                  )}
+                </div>
               </div>
 
               <div className={styles.check}>
-                <label>
-                  <input
-                    type="checkbox"
-                    name="consent"
-                    checked={isChecked}
-                    onChange={handleChange}
-                  />
-                  By clicking submit, I consent to my personal information being
-                  processed by Touchkin to address my request
+                <div className={styles.checked}>
+                  <label>
+                    <input
+                      type="checkbox"
+                      name="consent"
+                      checked={isChecked}
+                      onChange={(e) => setIsChecked(e.target.checked)}
+                    />
+                    &nbsp; By clicking submit, I consent to my personal information being
+                    processed by {WEBSITE_NAME} to address my request. &nbsp;
                 </label>
+                {error && !formData.isChecked && (
+                  <span className={styles.text_danger}>You must consent to the processing of your personal information</span>
+                )}
               </div>
-
-              {error && <div className={styles.error}>{error}</div>}
-
-              <div className={styles.btns}>
-                <Button btnClick={handleSubmit} size="16px" radius="5px" btnText="Sign Up" />
-              </div>
-            </form>
           </div>
-        </div>
-      )}
+
+          {error && <div className={styles.error}>{error}</div>}
+
+          <div className={styles.btns}>
+            <Button
+              btnClick={handleSubmit}
+              size="16px"
+              radius="5px"
+              btnText="Sign Up"
+            />
+          </div>
+        </form>
+
+          </div >
+        </div >
+      )
+}
     </>
   );
 }
 
 export default Signup;
-
